refactor:split wallet name settings
